backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
qdrant-client==1.7.0
sentence-transformers==2.3.1
huggingface-hub==0.20.0
langchain==0.1.0
langchain-community==0.0.10
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
